- Reorganize file structure
