+++ examples/Demo/FluentUI.Demo.Client/Documentation/Components/Theme/Designer/PaletteBar.razor
