+++ examples/Demo/FluentUI.Demo.Client/Documentation/Components/Theme/Designer/PaletteItem.razor
